# runtime/internal/openapidocs/openapi.json
{
  "components": {
    "schemas": {
      "APIError": {
        "properties": {
          "error": {
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "type": {
                "type": "string"
              }
            },
            "type": "object"
          }
        },
        "type": "object"
      },
...
      "llmretry_RetryPolicy": {
        "properties": {
          "fallback_after": {
            "type": "integer"
          },
          "fallback_model_id": {
            "type": "string"
          },
          "initial_backoff": {
            "description": "nanoseconds",
            "type": "integer"
          },
          "jitter": {
            "type": "number"
          },
          "max_attempts": {
            "type": "integer"
          },
          "max_backoff": {
            "description": "nanoseconds",
            "type": "integer"
          },
          "rate_limit_min_wait": {
            "description": "nanoseconds",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "taskengine_LLMExecutionConfig": {
        "properties": {
          "hide_tools": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "max_tokens": {
            "type": "integer"
          },
          "model": {
            "type": "string"
          },
          "models": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "pass_clients_tools": {
            "type": "boolean"
          },
          "provider": {
            "type": "string"
          },
          "providers": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "retry_policy": {
            "$ref": "#/components/schemas/llmretry_RetryPolicy"
          },
          "shift": {
            "type": "boolean"
          },
          "temperature": {
            "type": "number"
          },
          "think": {
            "type": "string"
          },
          "tools": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "tools_policies": {
            "additionalProperties": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "type": "object"
          }
        },
        "type": "object"
      },
      "taskengine_TaskChainDefinition": {
        "properties": {
          "debug": {
            "type": "boolean"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/taskengine_TaskDefinition"
            },
            "type": "array"
          },
          "token_limit": {
            "type": "integer"
          }
        },
        "type": "object"
      },
      "taskengine_TaskDefinition": {
        "properties": {
          "description": {
            "type": "string"
          },
          "execute_config": {
            "$ref": "#/components/schemas/taskengine_LLMExecutionConfig"
          },
          "handler": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "input_max_bytes": {
            "type": "integer"
          },
          "input_var": {
            "type": "string"
          },
          "output_template": {
            "type": "string"
          },
          "print": {
            "type": "string"
          },
          "prompt_template": {
            "type": "string"
          },
          "retry_on_failure": {
            "type": "integer"
          },
          "system_instruction": {
            "type": "string"
          },
          "timeout": {
            "type": "string"
          },
          "tools": {
            "$ref": "#/components/schemas/taskengine_ToolsCall"
          },
          "transition": {
            "$ref": "#/components/schemas/taskengine_TaskTransition"
          }
        },
        "type": "object"
      },
      "taskengine_TaskTransition": {
        "properties": {
          "branches": {
            "items": {
              "$ref": "#/components/schemas/taskengine_TransitionBranch"
            },
            "type": "array"
          },
          "on_failure": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "taskengine_ToolsCall": {
        "properties": {
          "args": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object"
          },
          "name": {
            "type": "string"
          },
          "tool_name": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "taskengine_TransitionBranch": {
        "properties": {
          "edge": {
            "type": "string"
          },
          "goto": {
            "type": "string"
          },
          "operator": {
            "type": "string"
          },
          "when": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "HTTP API served by `contenox serve` under /api. Generated from route annotations by tools/openapi-gen.",
    "title": "Contenox Runtime API",
    "version": "v0.36.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/taskchains": {
      "delete": {
        "operationId": "taskchain_deleteTaskChain",
        "parameters": [
          {
            "description": "Relative chain JSON path inside .contenox.",
            "in": "query",
            "name": "path",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "deleteTaskChain removes a task-chain definition file by path.",
        "tags": [
          "taskchain"
        ]
      },
      "get": {
        "operationId": "taskchain_getTaskChain",
        "parameters": [
          {
            "description": "Relative chain JSON path inside .contenox.",
            "in": "query",
            "name": "path",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/taskengine_TaskChainDefinition"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "getTaskChain reads one task-chain definition from its file path (query parameter `path`).",
        "tags": [
          "taskchain"
        ]
      },
      "post": {
        "operationId": "taskchain_createTaskChain",
        "parameters": [
          {
            "description": "Relative chain JSON path inside .contenox.",
            "in": "query",
            "name": "path",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/taskengine_TaskChainDefinition"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/taskengine_TaskChainDefinition"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "createTaskChain writes a new task-chain definition file and returns the stored definition.",
        "tags": [
          "taskchain"
        ]
      },
      "put": {
        "operationId": "taskchain_updateTaskChain",
        "parameters": [
          {
            "description": "Relative chain JSON path inside .contenox.",
            "in": "query",
            "name": "path",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/taskengine_TaskChainDefinition"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/taskengine_TaskChainDefinition"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "updateTaskChain overwrites an existing task-chain definition file and returns the stored definition.",
        "tags": [
          "taskchain"
        ]
      }
    },
    "/taskchains/list": {
      "get": {
        "operationId": "taskchain_listTaskChains",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "listTaskChains returns the relative file paths of every stored task-chain definition.",
        "tags": [
          "taskchain"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/api"
    }
  ]
}